Update README.md
This FIT-ME management system is an easy way to use gym and health membership system.
It can help to keep the records of registered members, guidance which exercise and muscle
groups to work out together, how much weight loss is required, their diet plans, logs of calories,
daily targets to achieve. So, keeping in mind the main features of FIT-ME are guidance to their
members about the
 Health and fitness checklist
 Weekly/Monthly Diet plan
 Daily Exercise chart
 Weight gain/loss plans
 Nutrient/Calories intake logs
 and many more
FIT-ME should be designed to facilitate a workout application to automate its operations of
keeping records and store them in the form of a large and user-friendly database further
facilitating easy access to the personnel.
The objective of the FIT-ME is to provide a system which handles the information of the
members using the application and maintaining their health care.